netlify/functions/requirements.txt
Flask
requests
python-dotenv
serverless-wsgi